Refactor component imports and add global styles for theming
